Refactor DeckPreviewWidgets to reside in their own folder.
