+++ cockatrice/CMakeLists.txt
@@ -161,10 +161,9 @@ set(cockatrice_SOURCES
     src/game/zones/view_zone_widget.cpp
     src/game/zones/view_zone.cpp
     src/client/tabs/visual_deck_storage/tab_deck_storage_visual.cpp
-    src/client/ui/widgets/cards/deck_preview_card_picture_widget.cpp
-    src/client/ui/widgets/visual_deck_storage/deck_preview_color_identity_widget.cpp
-    src/client/ui/widgets/visual_deck_storage/deck_preview_color_identity_filter_widget.cpp
-    src/client/ui/widgets/visual_deck_storage/deck_preview_widget.cpp
+    src/client/ui/widgets/cards/deck_preview_card_picture_widget.cpp src/client/ui/widgets/visual_deck_storage/deck_preview/deck_preview_color_identity_widget.cpp
+    src/client/ui/widgets/visual_deck_storage/deck_preview/deck_preview_color_identity_filter_widget.cpp
+    src/client/ui/widgets/visual_deck_storage/deck_preview/deck_preview_widget.cpp
     src/client/ui/widgets/visual_deck_storage/visual_deck_storage_widget.cpp
     src/client/ui/widgets/visual_deck_storage/visual_deck_storage_search_widget.cpp
     src/client/ui/widgets/visual_deck_storage/visual_deck_storage_sort_widget.cpp
