+++ cockatrice/src/client/ui/widgets/visual_deck_storage/deck_preview/deck_preview_color_identity_filter_widget.h
@@ -1,7 +1,7 @@
 #ifndef DECK_PREVIEW_COLOR_IDENTITY_FILTER_WIDGET_H
 #define DECK_PREVIEW_COLOR_IDENTITY_FILTER_WIDGET_H
 
-#include "visual_deck_storage_widget.h"
+#include "../visual_deck_storage_widget.h"
 
 #include <QChar>
 #include <QHBoxLayout>
