+++ cockatrice/src/client/ui/widgets/visual_deck_storage/deck_preview/deck_preview_color_identity_widget.cpp
@@ -1,6 +1,6 @@
 #include "deck_preview_color_identity_widget.h"
 
-#include "../../../../settings/cache_settings.h"
+#include "../../../../../settings/cache_settings.h"
 
 #include <QPainter>
 #include <QResizeEvent>
