+++ cockatrice/src/client/ui/widgets/visual_deck_storage/deck_preview/deck_preview_widget.cpp
@@ -1,7 +1,7 @@
 #include "deck_preview_widget.h"
 
-#include "../../../../game/cards/card_database_manager.h"
-#include "../cards/deck_preview_card_picture_widget.h"
+#include "../../../../../game/cards/card_database_manager.h"
+#include "../../cards/deck_preview_card_picture_widget.h"
 
 #include <QFileInfo>
 #include <QMouseEvent>
