+++ cockatrice/src/client/ui/widgets/visual_deck_storage/visual_deck_storage_widget.h
@@ -5,8 +5,8 @@
 #include "../../../../deck/deck_view.h"
 #include "../../../ui/widgets/general/layout_containers/flow_widget.h"
 #include "../cards/card_size_widget.h"
-#include "deck_preview_color_identity_filter_widget.h"
-#include "deck_preview_widget.h"
+#include "deck_preview/deck_preview_color_identity_filter_widget.h"
+#include "deck_preview/deck_preview_widget.h"
 #include "visual_deck_storage_search_widget.h"
 #include "visual_deck_storage_sort_widget.h"
